Rework films reducer with slice
